Make `vlag.py` pytest-ready
